+++ input/postgres/relation_stats_ext.go
@@ -63,9 +63,9 @@ func GetRelationStatsExtended(ctx context.Context, logger *util.Logger, db *sql.
 				server.SelfTest.MarkDbCollectionAspectError(dbName, state.CollectionAspectExtendedStats, "monitoring helper function pganalyze.get_relation_stats_ext not found")
 				server.SelfTest.HintDbCollectionAspect(dbName, state.CollectionAspectExtendedStats, "Please set up"+
 					" the monitoring helper function pganalyze.get_relation_stats_ext (%s)"+
-					" or connect as superuser, to get extended statistics for all tables.", selftest.URLPrinter.Sprint("https://github.com/pganalyze/collector#setting-up-a-restricted-monitoring-user"))
+					" or connect as superuser, to get extended statistics for all tables.", selftest.URLPrinter.Sprint("https://pganalyze.com/docs/install/troubleshooting/ext_stats_helper"))
 				logger.PrintInfo("Warning: Limited access to extended table statistics detected in database %s. Please set up"+
-					" the monitoring helper function pganalyze.get_relation_stats_ext (https://github.com/pganalyze/collector#setting-up-a-restricted-monitoring-user)"+
+					" the monitoring helper function pganalyze.get_relation_stats_ext (https://pganalyze.com/docs/install/troubleshooting/ext_stats_helper)"+
 					" or connect as superuser, to get extended statistics for all tables.", dbName)
 			}
 		}